Fix: improvements on express relay sdks and simple searchers (#2045)
* feat: Cache mint decimals and use global config from the order

* Improve svm types

* Minor updates on ts package
