+++ sdk/python/README.md
@@ -16,10 +16,28 @@ $ poetry add express-relay
 
 ## Quickstart
 
-To run the simple searcher script, navigate to `python/` and run
+To run the simple searcher script, navigate to `python/` and run the following command:
+
+### Evm
 
 ```
-$ poetry run python3 -m express_relay.searcher.examples.simple_searcher --private-key <PRIVATE_KEY_HEX_STRING> --chain-id development --verbose --server-url https://per-staging.dourolabs.app/
+$ poetry run python3 -m express_relay.searcher.examples.simple_searcher_evm \
+--private-key <PRIVATE_KEY_HEX_STRING> \
+--chain-id development \
+--verbose \
+--server-url https://per-staging.dourolabs.app/
 ```
 
-This simple example runs a searcher that queries the Express Relay liquidation server for available liquidation opportunities and naively submits a bid on each available opportunity.
+This simple example runs a searcher that queries the Express Relay liquidation server for available liquidation
+opportunities and naively submits a bid on each available opportunity.
+
+### Svm
+
+```
+$ poetry run python3 -m express_relay.searcher.examples.simple_searcher_svm \
+--endpoint-express-relay https://per-staging.dourolabs.app/ \
+--chain-id development-solana \
+--private-key-json-file <PATH_TO_JSON_FILE> \
+--endpoint-svm https://api.mainnet-beta.solana.com \
+--bid 10000000 # Bid amount in lamports
+```
