+++ sdk/python/express_relay/models/svm.py
@@ -1,4 +1,5 @@
 import base64
+from datetime import datetime
 from typing import Any, Annotated, ClassVar
 
 from pydantic import (
@@ -6,20 +7,23 @@ from pydantic import (
     GetJsonSchemaHandler,
     BaseModel,
     model_validator,
+    Field,
+    Base64Bytes,
 )
 from pydantic.json_schema import JsonSchemaValue
 from pydantic_core import core_schema
 from solders.hash import Hash as _SvmHash
-from solders.signature import Signature as _SvmSignature
 from solders.pubkey import Pubkey as _SvmAddress
+from solders.signature import Signature as _SvmSignature
 from solders.transaction import Transaction as _SvmTransaction
 
-from express_relay.express_relay_types import (
+from express_relay.svm.generated.limo.accounts.order import Order
+from express_relay.models.base import (
     IntString,
     UUIDString,
     UnsupportedOpportunityVersionException,
+    BidStatus,
 )
-from express_relay.svm.generated.limo.accounts import Order
 
 
 class _TransactionPydanticAnnotation:
@@ -180,6 +184,17 @@ SvmHash = Annotated[_SvmHash, _HashPydanticAnnotation]
 SvmSignature = Annotated[_SvmSignature, _SignaturePydanticAnnotation]
 
 
+class BidSvm(BaseModel):
+    """
+    Attributes:
+        transaction: The transaction including the bid
+        chain_id: The chain ID to bid on.
+    """
+
+    transaction: SvmTransaction
+    chain_id: str
+
+
 class _OrderPydanticAnnotation:
     @classmethod
     def __get_pydantic_core_schema__(
@@ -228,7 +243,7 @@ class OpportunitySvm(BaseModel):
         version: The version of the opportunity.
         creation_time: The creation time of the opportunity.
         opportunity_id: The ID of the opportunity.
-        blockHash: The block hash to use for execution.
+        block_hash: The block hash to use for execution.
         slot: The slot where this order was created or updated
         program: The program which handles this opportunity
         order: The order to be executed.
@@ -240,7 +255,7 @@ class OpportunitySvm(BaseModel):
     creation_time: IntString
     opportunity_id: UUIDString
 
-    blockHash: SvmHash
+    block_hash: SvmHash
     slot: int
 
     program: str
@@ -258,3 +273,45 @@ class OpportunitySvm(BaseModel):
                 f"Cannot handle opportunity version: {data['version']}. Please upgrade your client."
             )
         return data
+
+
+class BidStatusSvm(BaseModel):
+    """
+    Attributes:
+        type: The current status of the bid.
+        result: The result of the bid: a transaction hash if the status is SUBMITTED or WON.
+                The LOST status may have a result.
+    """
+
+    type: BidStatus
+    result: SvmSignature | None = Field(default=None)
+
+    @model_validator(mode="after")
+    def check_result(self):
+        if self.type == BidStatus.WON or self.type == BidStatus.SUBMITTED:
+            assert (
+                self.result is not None
+            ), "bid result should not be empty when status is won or submitted"
+        return self
+
+
+class BidResponseSvm(BaseModel):
+    """
+    Attributes:
+        id: The unique id for bid.
+        bid_amount: The amount of the bid in lamports.
+        chain_id: The chain ID to bid on.
+        permission_key: The permission key to bid on.
+        status: The latest status for bid.
+        initiation_time: The time server received the bid formatted in rfc3339.
+        profile_id: The profile id for the bid owner.
+    """
+
+    id: UUIDString
+    bid_amount: int
+    chain_id: str
+    permission_key: Base64Bytes
+    status: BidStatusSvm
+    initiation_time: datetime
+    transaction: SvmTransaction
+    profile_id: str | None = Field(default=None)
